requirements.txt
openai>=1.12.0
chromadb>=0.5.0   # or another vector DB, choose one and stick with it
pypdf>=4.0.0
python-dotenv>=1.0.0

